Refactor code reproducing ESP field for modularity
The field_func, which dictates what property of the field is to be
evaluated at each point, is now separate from the method iterating the
grid and calling field_func at each point. This increases the modularity
of the code to easily allow new field_funcs. The code was refactored
with a _dist_func in mind, which will return two values, hence the new
method returns as many fields as the field_func returns values.

The code is kludgey in terms of how the field_types are found. This will
be fixed in the future.