Add NonGridField object to keep values at coordinates
The `G09_esp` object, which is more of an interface to the `g09` `.esp`
input file, now creates a `NonGridField` object. This is a simple
object, which keeps the coordinates, at which the fitting was performed,
and the corresponding ESP values. It also performs more thorough input
format checks than the `G09_esp` object, including a check for duplicate
points. The method for writing an .esp file in a format expected by
`resp` has also been moved to this new object. Most importantly,
however, it is necessary to contain some helpers for making it work like
a `Field` object (ESP reproduction, distance transform etc.), which will
be a plenty of work.