Create a Field object as parent of old Field
The old Field has been renamed to GridField and inherits from the new
Field. NonGridField also inherits from the new Field. This is done to
keep the common methods and interface in the common Field, but both
children will have their own specifics, where their behaviour depends on
the presence of the grid.