Separate NonGridField from its Points
Three out of four parts of the data associated with a NonGridField are
the coordinates of its points in 3D. These coordinates might be
associated with many fields, so must be separate from the values. This
is in analogy to GridField, which uses a separate object for its grid
(even though the grid itself is lightweight).