Move creation of Points object out of NonGridField
The `Points` object needs to be created separately, e.g. by the
`G09_esp` object, which is one of the ways that a `NonGridField` can be
created.  This is analogous to how a `Grid` object is created by a
`Cube` and passed as one of the parameters to initialization of a
`GridField`. This allows to manipulate `Grid` and `Points` objects more
independently, which is especially important for the latter, which is
rather big.

This change allowed removing some of the initialization parameters of
`NonGridField`. The responsibility for handling duplicates and units is
now entirely with the `Points` object.

This change also resolves a minor imprecision in module `rep_esp`.
Previously the function `calc_non_grid_field` initialized NonGridField
with default values of its keyword arguments, which suggested that the
points were being passed in Bohr radii and would be converted. Since a
Points object was being passed, no conversion was taking place, and the
passed distances in Angstroms were rightly not further converted anyway.