Move parameter type checking between functions
This change does not affect the code as it is now, since the only other
function (`calc_grid_field`) that would call `_calc_field` creates the
object being checked (`Points`) itself. However, having the check in
this more fundamental function makes more sense in case it is called
from elsewhere and also serves as a verification if any changes were
made to `calc_grid_field`.