Extract QTAIM charges from Henkelman's `bader`
The added `.dat` files for methane have been calculated with a cube of a
very fine grid (500 points on each side). The charges were not
consistent with those from AIMAll (`.sumviz` file), so a warning is
issued when this method is used. Also, note that implementing this in
the code wasn't very clean, because the basin charges had to be
subtracted from nuclear charges to yield net charge.