Rectify the issue with fixing shell thickness
I realized that IOp43 ('increment between shells') likely refers to
multiples of van der Waals radii, just like in the MK paper, rather than
a value in angstroms. However, investigating the distances on my 2D plot
of methane showed, that the value used for hydrogen is unusual. I did
some research and described it extensively in the docstrings. A formula
in the code had to be updated to account for the vdW radius factor in
the increment.