Change default Sphinx settings
Create the documentation by running from the `docs` directory:

sphinx-apidoc -fMe -o source ../repESP && make html

See: http://stackoverflow.com/a/35250460

The documentation is really not up to scratch now, this is just a bit of
a skeleton to make me think about it and see if I'm formatting the
docstrings properly. The `numpy` docstring formatting will be used with
the help of the `napoleon` extension.