Add script plotting the fit for various charge ratios
It would be nice to be able to plot both regular and heavy ratio on the
same plot but that can't easily be done with the current setup. A
similar issue is present in the `esp_fit` script, so some
changes/refactoring will be probably be relevant to both these scripts.