Don't require all files when looking for respin & esp
Previously the filenames were optional, and setting them to `None` just
requested that they are found automatically. Now setting them to `None`
means that they should not be searched and returned. To search for them
automatically, an empty string should be passed, hence the default
values of `run_resp` had to be changed.