Major refactor: parameters at top of script
The script was refactored to make it less specific to the molecules for
which it was developed. While it is not completely general (it only
makes sense for certain molecules), all the parameters have been
gathered at the top of the script and can be tweaked without delving
into the code too much. Note that in the 2D contour presentation part,
`ivary_label2` corresponds to the horizontal axis. This is intentional
(in order to keep `ivary_label1` as C1, just like in methane, but have
N13 or N17 on the horizontal axis in contour plots) but doesn't quite
align with the least surprise rule, so should perhaps be changed in the
future to prevent bugs.