Enable running *both* 1 and 2D variations
Both 1 and 2D variations are now run by default.

This required some changes regarding the `vary_label` variables. Now the
first label refers to the primary atom of interest and a 1D scan will be
performed on it. The second label will be for the other atom used in the
2D scan. The charge dictionaries were also updated and the conceptual
reference to C and N atoms in variable names was removed (now they're
just `a1` and `a2`).

A significant portion of the code change in this commit, especially the
second half, is swapping the positions of vary_labels in 2D variation.
The reason is that previously the axes had to be swapped to yield a
reasonable-looking contour plot (due to the slant of the ellipses for
NMe3H). Now the user may still want to swap the axis for some other pair
of charges and will need to do it manually, which is annoying.