Change appearance of plots (markers and colourmap)
The colourmap was unfortunately changed from a colour-blind-friendly,
because `plasma` was found not intuitive enough. `coolwarm_r` is similar
to the one used in plotting the fitting points `RdYlBu`, but that one
had yellow neutral values which were not visible on this plot with the
new edge-less and smaller markers.