Create a new parent parser for RESP-related options
The --esp_file option will be a positional argument in the compromise
script, so its help was moved to the parent but the option is add to
parser in the averaging script.

The information about the esp_file being used for equivalencing is
specific to averaging and was hence moved to the help description of
that script.