Fix missing guiding line in response plot in script
Plotting the guiding line (x = 0) was moved from `plot_flexibility_func`
to code common to both plotting functions.

This issue was similar to the missing x-label from response plot and the
new code has to be independent of the data on the flexibility plot. This
required fetching the y-limits first and later restoring them back.