Simplify Coords from class to NewType
This was at the cost of the convenient constructor from three floats.
I'm not sure if it should've been there in the first place considering
it's a bit of a shortcut with respect to the types. A helper was defined
in tests, but that's understandable. Applications can do the same if
their usages turn out to be repetitive for some reason too.

Also removed all the custom comparisons, as I don't think the precision
should be decided by the library. Instead, applications will have the
choice of setting the right precision depending on the context.