Expose origin and axes of GridMesh
Private attributes don't really work with dataclasses, because __init__
expects them as keyword arguments, so a kwarg prefixed with an underscore
would have to be passed in. That's unless they're excluded from __init__
and handled differently.

I'm not convinced that the `Field` and `Mesh`es should be dataclasses
but they sure are convenient.