Remove redundant (due to single-structure) attrs from Respin
The Respin class can currently represent only a single-molecule
single-structure optimization, so the attributes were redundant and
there was a risk of the user changing the values, resulting in an
internally inconsistent dataclass. These are better represented as class
properties. The class will be renamed to make it clear that it can only
refer to a single structure in a commit to follow.