gaussian_format: Rethrow error parsing charges line as InputFormatError
It's not clear whether Gaussian output generated without the IOp(6/50=1)
override would be parsed successfully. This change makes it clear that
the problem is with the input format rather than with the library
itself. It hasn't yet been decided whether the library will be adapted
or if it will be documented that the the override is required.

This change also removes a comment in the code describing the possible
issue in favour of the issue on GitHub, https://github.com/jszopi/repESP/issues/28