Relax ctor interface of Coords to match that of tuple (#38)
Expecting `Tuple[Any, Any, Any]` was too strict, because it didn't match the interface that functions could expect from an object inheriting from a `tuple`. This resulted in many `mypy` errors that would be annoying to address. `__init__` now takes an `Iterable` and the check for the length has been moved to runtime.